+++ apps/plugins/lib/strncpy.c
@@ -41,13 +41,14 @@ QUICKREF
 #include <limits.h>
 #include "plugin.h"
 #include "_ansi.h"
+#include <stdint.h>
 
 /*SUPPRESS 560*/
 /*SUPPRESS 530*/
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define ROCKBOX_UNALIGNED(X, Y) \
-  (((long)X & (sizeof (long) - 1)) | ((long)Y & (sizeof (long) - 1)))
+  (((uintptr_t)X & (sizeof (long) - 1)) | ((uintptr_t)Y & (sizeof (long) - 1)))
 
 #if LONG_MAX == 2147483647L
 #define DETECTNULL(X) (((X) - 0x01010101) & ~(X) & 0x80808080)
