+++ firmware/asm/memset16.c
@@ -20,9 +20,10 @@
  ****************************************************************************/
 
 #include "string-extra.h" /* memset16() */
+#include <stdint.h>
 
 #define LBLOCKSIZE (sizeof(long)/2)
-#define ROCKBOX_UNALIGNED(X)   ((long)X & (sizeof(long) - 1))
+#define ROCKBOX_UNALIGNED(X)   ((uintptr_t)X & (sizeof(long) - 1))
 #define TOO_SMALL(LEN) ((LEN) < LBLOCKSIZE)
 
 void memset16(void *dst, int val, size_t len)
